useRouteLoaderData for user context
